Move client commands and parsers into a client subpackage
