+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -31,7 +31,8 @@ import static seedu.address.testutil.TypicalClients.BOB;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.AddCommand;
+import seedu.address.logic.commands.client.AddCommand;
+import seedu.address.logic.parser.client.AddCommandParser;
 import seedu.address.model.client.Address;
 import seedu.address.model.client.Client;
 import seedu.address.model.client.Email;
