+++ src/test/java/seedu/address/logic/parser/EditCommandParserTest.java
@@ -34,7 +34,8 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_THIRD_CLIENT;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.EditCommand;
+import seedu.address.logic.commands.client.EditCommand;
+import seedu.address.logic.parser.client.EditCommandParser;
 import seedu.address.model.client.Address;
 import seedu.address.model.client.Email;
 import seedu.address.model.client.Name;
