+++ src/test/java/seedu/address/logic/parser/FindCommandParserTest.java
@@ -8,7 +8,8 @@ import java.util.Arrays;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.FindCommand;
+import seedu.address.logic.commands.client.FindCommand;
+import seedu.address.logic.parser.client.FindCommandParser;
 import seedu.address.model.client.NameContainsKeywordsPredicate;
 
 public class FindCommandParserTest {
