+++ src/test/java/seedu/address/testutil/ClientUtil.java
@@ -8,8 +8,8 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
 import java.util.Set;
 
-import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.EditCommand;
+import seedu.address.logic.commands.client.AddCommand;
+import seedu.address.logic.commands.client.EditCommand;
 import seedu.address.model.client.Client;
 import seedu.address.model.tag.Tag;
